Add loop counter logic and AnimatableImageView delegate
The main loop logic was extracted from
https://github.com/kaishin/Gifu/pull/42, a open PR from @storix
that was still not merged at the time this commit was created

Add loop counter logic and optional delegate method which is called when
an animation stops

Add a public getter for the current loop count and a setter/getter for
the max number of loops. If max number of loops is set to 0 it loops
forever

Add loadedData which hold the data that is currently loaded in the
AnimatableImageView. It has a private setter and a public getter